requirements.txt
africastalking==1.2.9
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
azure-ai-inference==1.0.0b9
azure-core==1.32.0
beautifulsoup4==4.12.3
certifi==2025.1.31
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
confusable-homoglyphs==3.3.1
contourpy==1.3.0
cycler==0.12.1
distlib==0.3.9
dj-database-url==2.3.0
Django==4.2
django-client-side-image-cropping==0.1.10
django-cors-headers==4.6.0
django-heroku==0.3.1
django-registration==5.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
filelock==3.17.0
fonttools==4.55.4
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
importlib_resources==6.5.2
isodate==0.7.2
kiwisolver==1.4.7
matplotlib==3.9.4
multidict==6.2.0
mysql-connector-python==9.2.0
mysql-to-sqlite3==2.3.0
mysqlclient==2.2.7
numpy==2.0.2
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
pipenv==2024.4.1
platformdirs==4.3.6
propcache==0.3.1
psycopg2==2.9.10
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytimeparse2==1.7.1
requests==2.32.3
schema==0.7.7
simplejson==3.19.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tabulate==0.9.0
text-unidecode==1.3
tqdm==4.67.1
twilio==9.5.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
virtualenv==20.30.0
whitenoise==6.8.2
yarl==1.18.3
zipp==3.21.0
